Updated random seed
Using the same seed causes options to be shuffled the same way. This is intended so that every time a person gets the same question, the options are in the same order but might be different from another person's option order(aka randomised). However, I have forgotten that the options would be arranged the same way across questions, making the answer always the same version. Hence, I have changed the seed to include the username, question and question ID to generate the same options for a user for a question but different for different users and different questions.